+++ jmx/jmx_test.go
@@ -60,24 +60,6 @@ func TestMain(m *testing.M) {
 	}
 }
 
-func TestOpen_OnlyWorksWhenClosed(t *testing.T) {
-	defer Close()
-
-	assert.NoError(t, Open("", "", "", ""))
-	assert.Error(t, Open("", "", "", ""))
-	Close()
-	assert.NoError(t, Open("", "", "", ""))
-}
-
-func TestOpenWithSSL_OnlyWorksWhenClosed(t *testing.T) {
-	defer Close()
-
-	assert.NoError(t, OpenWithSSL("", "", "", "", "", "", "", ""))
-	assert.Error(t, OpenWithSSL("", "", "", "", "", "", "", ""))
-	Close()
-	assert.NoError(t, OpenWithSSL("", "", "", "", "", "", "", ""))
-}
-
 func TestOpenWithParameters_OnlyWorksWhenClosed(t *testing.T) {
 	defer Close()
 
@@ -169,7 +151,8 @@ func openWait(hostname, port, username, password string, attempts int) error {
 }
 
 func openWaitWithSSL(hostname, port, username, password, keyStore, keyStorePassword, trustStore, trustStorePassword string, attempts int) error {
-	err := OpenWithSSL(hostname, port, username, password, keyStore, keyStorePassword, trustStore, trustStorePassword)
+	ssl := WithSSL(keyStore, keyStorePassword, trustStore, trustStorePassword)
+	err := OpenWithParameters(hostname, port, username, password, ssl)
 	if err == ErrJmxCmdRunning && attempts > 0 {
 		attempts--
 		time.Sleep(10 * time.Millisecond)